Bug#52081: Index hint/forcing for GROUP BY / ORDER BY is broken
Found during code inspection: There is a mixup between GROUP BY 
and ORDER BY in which indexes that should be considered/allowed 
during optimization of a query. The bug is a simple mixup of 
argument 'is_order_by' to create_sort_index() in JOIN::exec().
      
No test case included as only problem is increased execution
time due to wrong plan, and this does not show in EXPLAIN.