This commit replaces a fixed size buffer for receiver ids with an iterator. This buffer/iterator is used when asking for a new batch of a scan. The fixed size array could potentially overflow for clusters with many nodes and replicas.
This commit should also fix a problem that causes the regression test for 
'ndbinfo.counters' for the SPJ block to fail on big endian architectures. 
Ndb uses an endian-dependent hashing algorithm. The number of lookup that are
excecuted on the locally versus remotly will thus also depend on endian-ness.
The regression test will now only test the total number of both local and remote
operations, which should always be the same.