Bug#46548 IN-subqueries return 0 rows with materialization=on
(Backporting of revid:oystein.grovlen@sun.com-20091123095520-ponntgejv3mju88o)

This bug is related to key lookups in a materialized table.  A caching 
mechanism is used to check whether you are doing lookup on the same key 
as the last time.  In that case, the right record is assumed to already 
be present in the record buffer (see cmp_buffer_with_ref()).

The problem is that it derives the key_length from the 
key_info of the materialized table.  However, when creating the table,
this key_length did not take into account extra bytes used to store null
information and length of variable length fields. When the key_length was
not correct, cmp_buffer_with_ref() would not compare the entire key, 
and there was a risk that it would wrongly assume that it already had 
the right record in the buffer.
 
The problem is fixed by making key_length of key_info always be the sum of
the store_length of its key_parts (in create_tmp_table()).  As it is for