Bug#46556 Returning incorrect, empty results for some IN subqueries w/ semijoin=on
(Backporting of revid:oystein.grovlen@sun.com-20091211140923-q0uu1djdd5de4ce0)

Problem: When a table is pulled out of a semijoin, it is not 
considered whether this makes it a correlated outer table. If 
it really is a correlated outer table, FirstMatch strategy
requires that it comes before the semijoin-nest in the join 
sequence.  Since this information is not recorded, this 
restriction is not enforced.

Solution: If a table is pulled out of semijoin nest due to 
eq_ref, add it to the set of tables that are regarded as 
correlated to the semijoin nest.