Bug#49630 "Segfault in select_describe() with double            nested subquery and materialization"
(Backporting of revid:jorgen.loland@sun.com-20100315075225-rcwlxcsuuidwnwzi)

If a JOIN contains const tables, make_join_select() will 
evaluate the conditions in the WHERE clause. If the WHERE 
clause contains a subquery, the process of evaluating 
the condition involves optimize() and exec(). 

Calling optimize() and exec() on a subselect may require the
use of temp tables, in which case the original JOIN query
execution plan will be replaced by a simple scan of the 
temp table. 

To be able to describe the query plan, the original query 
layout needs to be saved. This was not done for materialized
subqueries. The fix is to make materialized subqueries save 
the original join query layout if it is needed by EXPLAIN.