Refactoring of how parent/child tuples are correlated in the result sets:
Previously the 3 sub structures *m_parentTupleId, *m_childTupleIdx and m_tupIdToTupNumMap, 
was allocated as part of each NdbResultStream in order to handle correlation of parentId / childId 
in a linked scan-result set.
A Vector was also used as part of tupIdToTupNumMap, which could be reallocated multiple
times for a large BatchSize.

Furthernmore, buildChildTupleLinks() had a complexity of 0 * (BatchSize ^2) which caused 
population of m_childTupleIdx to slow down with larger 'BatchSize'.

This refactoring commit collects information from all these 3 substructures into the new
class TupleSet. A single array of TupleSet objects are allocated when the 
NdbResultStream is constructed, which eliminates all the 3 previous objects *and* the Vector.

A HashMap structure is also implemented on top of the TupleSet[] in order to speed up
location of correlated parent/child tuples. Lookup through this HashMap has made
buildChildTupleLinks() obsolete.