Fixed regression caused by http://lists.mysql.com/commits/104509:
NdbQuery::setErrorCode() will also  set the 'm_state' of 
the NdbQuery to 'Failed'. As the condition for NdbQuery::hasCompleted() 
use to be 'm_state >= Closed' ('Failed' *is* > 'Closed') a Failed NdbQuery
would be released by NdbTransaction::releaseCompletedQueries().

This caused later explicit calls to NdbQuery::close() to refer an
already destructed NdbQuery object.

The solution is to require the NdbQuery to have been 
::close'ed before it can be garbage collected by 
releaseCompletedQueries()

NOTE: A 'transaction end' will always force any open NdbQuery object to
be closed and garbage collected -> There can't be any leaked NdbQueryes 
past a transaction end.