BUG#44486: master_retry_count cannot be displayed
SHOW SLAVE STATUS did not show master_retry_count.

We fix this by adding a line to the end of the output
of SHOW SLAVE STATUS that prints out the current value
of master_retry_count.

Additionally, several test result files needed to be
updated (due to the extra output in SHOW SLAVE STATUS).