Bug #49198 Wrong result for second call of procedure with view in subselect
Backport of tor.didriksen@sun.com-20100208113127-g8wvxcpc2dppmxxd

The problem was that fix_after_pullout() after semijoin conversion 
wasn't propagated from the view to the underlying table. 
On subesequent executions of the prepared statement, 
we would mark the underlying table as 'dependent' and the predicate 
anlysis would lead to a different (and illegal) execution plan.