BUG#52166: Seconds_Behind_Master spikes after long idle period
Introduction: seconds behind master is calculated by checking the
time difference between the current time on the slave (time(0))
and the creation timestamp, at the master, of the last executed
event by the SQL thread. The last event timestamp is saved by the
SQL thread in rli->last_master_timestamp. Different timezones for
master and slave are taken into account when showing this
delta. The rli->last_master_timestamp is updated when the SQL
thread finishes the event execution at the slave.

Problem: when the SQL thread has no events to replay in the relay
log, it waits. Right before waiting, it saves the last
event (lets call it e1) timestamp (rli->last_master_timestamp)
set at the master (in save_last_master_timestamp), and sets
rli->last_master_timestamp to zero. This makes 'SHOW SLAVE
STATUS' show zero while the SQL thread is waiting. However, when
the SQL thread awakes, most likely because of the fact that a new
event was enqueued in the relay log, it restores the
save_last_master_timestamp to rli->last_master_timestamp. Then it
goes on to process the new event (lets name it e2). Given this
and the fact that the SQL thread only updates the
rli->last_master_timestamp at the end of execution, there is a
time window that the user can notice a sudden spike on
Seconds_Behind_Master, especially if e2 arrives a long time after
e1 finished. The spike is actually enforced - value increases -
during e2 execution time, which means that if e2 execution
latency increases, then the likelihood that the user sees this
also increases.

Solution: We fix this by not saving and restoring the
last_master_timestamp when the SQL thread goes to sleep. Instead,
we set the rli->last_master_timestamp to zero, when the SQL
thread starts waiting, and keep it to zero, when the SQL thread
awakes. A complementary change in this patch, is the one that
sets rli->last_master_timestamp before and not after event
execution. Both of these changes make the spike go away, and keep
showing zero, until the event starts being processed, as opposed
to showing zero until the event finishes its execution.