Bug #52381 Cannot connect to server with IPv6 using hostname or IPv4 address
The bug occurs on Visual Studio 2005. IPV6_V6ONLY TCP flag is not
defined in the VS2005 SDK (it is an older pre-Vista one).
The code in mysqld.cc responsible for creating dual-mode socket is
enclosed into #ifdef IPV6_V6ONLY. When compiled with VS2005,
v6only socket option is not cleared which results into server listening
only on IPV6 address.
      
There used to be a workaround for this problem, defining IPV6_V6ONLY expli
citly when it is not defined in system headers. The  system checks  for it 
are already in place in CMake, but the results were precached 
(i.e check was effectively deactivated). The patch reactivates checks for 
IPV6_V6ONLY on pre-VS2008 Visual Studio. If these constants are  not found, 
my_config.h will contain
#define IPV6_V6ONLY 27
and this will make mysqld.cc compile with socket dual-mode  again.