Bug#46792 Hard coded 'mysql' table schema in error messages
Prior to this fix, the following error messages:
- ER_COL_COUNT_DOESNT_MATCH_CORRUPTED
- ER_CANNOT_LOAD_FROM_TABLE
did hard code a table schema to mysql.

With this fix, new error messages have been introduced:
- ER_COL_COUNT_DOESNT_MATCH_CORRUPTED_V2
- ER_CANNOT_LOAD_FROM_TABLE_V2
that uses a generic "%s.%s" table name.

The old error were renamed to ER_OBSOLETE_... for clarity,
and to maintain binary compatibility.

This fix also changed PFS_check_intact::report_error,
to make sure that errors are reported in the server log.
Previously, the message was lost.