BUG#43535: last_io_error in show slave status is confusing
Although IO and SQL thread errors numbers and messages are reported in
SHOW SLAVE STATUS output, the correspondent timestamp at which they
have occurred is not. As such the user may be left wondering whenever
the error happened (NOTE: actually, in the error log, a timestamp
should be printed along with the error message, however in SHOW SLAVE
STATUS it is not currently).
      
Furthermore, at the time this bug was reported, there was no way for
the user to clean the error number and message reported.
      
This patch adds a timestamp, similar to the one found in the error
log, to the Last_*_Error columns. It appends it at the beginning of
the error message, as in (square brackes not included):

 "Last_IO_Error    YYMMDD hh:mm:ss [PLAIN OLD ERROR MESSAGE ...]"
     
The computation of the timestamp is added as a new method of the Error
class (update_timestamp), which in turn is an inner class of the
Slave_reporting_capability. This new method is called everytime
Slave_reporting_capability::report is invoked with log level set to
ERROR_LEVEL, thence capturing the timestamp of the error being
reported.
      
Starting from 5.1.37 (ie, when the fix for BUG 44270 was released),
the fields are cleared on RESET SLAVE. Furthermore, the test case
added in this patch checks that: (1) the error message(s) is not reset
on explicit stop slave after an error occurred; (2) the error
message(s) is reset when slave recovers from the error (ie, when the
DBA fixes the problem and resumes replication by issuing START
SLAVE). Moreover, clearing the new Timestamp fields is done in the
Error class clear() method, therefore when Last_*_Errno and
Last_*_Error are reset, Last_*_Error_Timestamp is gets reset also.
      
Finally, this patch accommodates the changes needed for existing MTR
server tests:
      
 - result files are updated due to the new _Timestamp columns
      
 - show_slave_status.inc include files are augmented with
   replace_regex to mask out the timestamp from result files
      
 - some tests are also changed to mask out the output for the
   new _Timestamp columns