Bug#53392 Tests: perfschema.query_cache fails
Before this fix, the test perfschema.query_cache could fail
with spurious failures, depending on which test was executed in the same
MTR run.

With this fix, the test perfschema.query_cache resets the server
status variables, to be independent of previous runs.

In particular, Qcache_hit now has a predictable value.