Bug#53394 Tests: perfschema.myisam_file_io fails
Before this fix, the test perfschema.myisam_file_io could fail
with spurious failures, depending on which test was executed in the same
MTR run.

With this fix, the test perfschema.myisam_file_io resets the server
status variables, to be independent of previous runs.

In particular, performance_schema_*_lost now have a predictable value.