Fixed a bug in NdbQueryImpl::closeTcCursor() where a finalBatchReceived() may be counted both by the receive handling code and inside ::closeTcCursor() causing 'assert(m_finalBatchFrags == getRootFragCount());' to be thrown at end of ::CloseTcCursor().
The fix on this problem is to clear the 'm_fullFrags' in order to throw any 
remaining results before close issues a ::sendClose().

In addition to this fix. there is a small refactor of code related to 
handling results from the FragStack:

- As ::top() and ::pop() more or less contained the same functionality,
  ::top() has been deprecated. and replaced with ::pop().

- Added ::size() to be used in the single place where we actually needed to
check for items in the stack without also popping results.

Did not add a testcase as problem is timing dependent 
and could not be consistently reproduced. It will materialize 
in RQG test after some 1.000 queries containing scalar subqueries 
where the query fails due to the scalar subquery returning more than 1 row.