BUG#47367 Crash in Name_resolution_context::process_error
When subquery is transformed into a semijoin, a new name resolution object
is assigned for each of the column items of the subquery. The error_processor
field of this object was never set, so any error in resolution during second
or later execution would cause a crash when this procedure was invoked.
  
mysql-test/r/subselect3.result
  New test case for bug#47367
mysql-test/r/subselect3_jcl6.result
  New test case for bug#47367
mysql-test/t/subselect3.test
  New test case for bug#47367
sql/item.cc
  Fully initialize the name resolution object when it is created.