BUG#44209: MASTER_CONNECT_RETRY and --master-retry-count disconnected            from each other
Post-push fix! One of the test cases is not robust enough because
it depends on a time variable. This can cause it to fail because
it is a source of non-determinism (especially on slow environments).

The specific test case is about waiting for slave to stop after
retrying the N expected times and then checking that no more than
a number of expected seconds has ellapsed. This assertion is not
deterministic.

We fix the assertion by checking the number of retries reported by
the server, instead of checking how much seconds have ellapsed.