WL#4783  Extend mysqlbinlog to support remote backup of binary logs
The aim is to modify mysqlbinlog so that it can remotely retrieve the raw 
binary logs and stay connected retrieving them.  This way you can have an 
up-to-the-second remote backup of your data (even better when combined 
with mysqldump --master-data as snapshots).
            
You retrieve from many servers simultaneously into a backup server using this 
method (use a mysqlbinlog instance per server).
           
Changes:
* mysqlbinlog now safely exits with CTRL-C or kill (closes files properly)
* --raw added which will retrieve raw binlog files from the remote server and 
  save them using the filenames in the replication headers
* --stop-never added which waits for new data on a remote connection 
  instead of quitting at the end of the last log
* --stop-never-slave-server-id added to set the MySQL slave server ID for 
  --stop-never (default 65535)
* --result-file modified so that when used with --raw mode it is a prefix for 
  the output files