Bug#45191: Incorrectly initialized semi-join led to a wrong result.
Current implementation of join cache allows combination with the FirstMatch
semi-join strategy only when FirstMatch applied only to one table.
But because of a typo in the FirstMatch initialization code, all tables in
the semi-join except the first one was incorrectly marked as not belonging
to the semi-join. This wrongly allowed join cached usage and led to a wrong
result.

The setup_semijoin_dups_elimination function now correctly marks all tables
of a semi-join when the FirstMatch strategy is used.
Origianl revid: epotemkin@mysql.com-20090901144435-7n2doi4s8yn5hz3l