Bug #48868 Left outer join in subquery causes segmentation fault in make_join_select
Backport of tor.didriksen@sun.com-20100108093836-405e9flc3x1egvbt

This problem with left outer join in IN-subqueries that surfaced
when the left table was pulled out because it was constant (one row).
The problem was related to a loop at the end of make_join_select
              
  for (tab= join->join_tab+join->const_tables; tab <= last_tab ; tab++)
              
The problem was that the 'tab' loop variable was also the loop
variable of an outer loop.  There was code below this inner loop
that used tab that gets into trouble because of this.  The inner
loop was only used for inner tables of outer joins.  Hence the
problem only surfaced for outer joins.

Solution: split make_join_select() in two, and put the
condition-push-down code into a separate utility function.