Bug#53617 Missing performance schema tables not reported in the server log at startup
Before this fix, when the performance schema tables are missing due to a broken
installation or upgrade, the server would start but not report missing performance schema tables.

With this fix, missing tables are reported in the server error log, like for example as:
ERROR Native table 'performance_schema'.'SETUP_TIMERS' has the wrong structure

This test can not be automated: tested manually, no script provided.