BUG#53643: assert in Field_new_decimal::store_value on slave server
The table->write_set at the slave was improperly set before
processing a row event. For update events, it was being set
to the read set shipped in the event instead of the write_set.

This would lead to an assertion while unpacking the event 
because while setting some fields the corresponding bit in
the slave's table->write_set was not set.

We fix this by properly initializing the read_set and write_set
at unpacking time.