Prepares DBSPJ to also accept pushed joins with lookup as root *and* with a scanfilter (or 'InterpretedCode') on the root operation.
For consistency we require (and asserts) that both the request operation 
and the treeNode of the root operation contains the same information
wrt. getInterpretedFlag().

A similar assert already existed for getNoDiskFlag() - However, this assert
was not restrictive enough as it was sufficient that either the request or the
treeNode had the NoDiskFlag. For code consistency this assert was also 
changed to follow the same pattern as getInterpretedFlag() above.

Also:
- Simplified a consistency check for NI_KEY_PARAMS / PI_KEY_PARAMS.
- Added a missing setting of T_ATTR_INTERPRETED.