WL#5092: fix for perfschema binlog_mix(row) failing test(s).
First, replicating performance schema tables is meaningless.
Second, there are implicit checks that make certain fields in
perfschema tables to be readonly. If we try to update these
fields an error is reported.

Therefore, when using full RBR rows (which is the default), the
readonly fields will be signalled in the write_set, which will
ultimately result in an error.

To circumvent this we could either: 1. ignore perfschema tables
when setting the table->write_set according to the
binlog-row-image option, or 2. make a handlerton
flag (HA_NO_BINLOG_ROW_OPT - handler does not comply with RBL 
optimizations), that can be exported by perfschema handlerton, 
so that binlog-row-image procedure applies conditionally, based 
on whether the table exports the flag or not.

We went for option #2 because it does not enforce an explicit
reference to perfschema in the server and also because it can be
seamlessly reused by other engines that wish to achieve the same
effect. Additionally, other engines can benefit from this approach
in the future for similar scenarios.