wl#2540 replication checksum
intermediate changeset implements the task w/o relying on FD (to be refined by the following patch) as well as with per-event (A) (should be removed from all but FD). The 3rd todo will be correct affected tests because of FD is going to be extended by (A) size of 1 bytte. Finally, merging with fixes for bug#49741 shall complete the show