WL#5096: Write test cases for WL#5092
The rpl_row_img is a large test case. It was sporadically 
timing out on sol10 platform (because it would take too long
to execute). To work around this, we split the test case into
three logical units: engines, indexes and blob scenarios,
therefore the original rpl_row_img is unfolded into:
  - rpl_row_img_engines
  - rpl_row_img_indexes
  - rpl_row_img_blobs

Still a big test, but done in smaller units so that we work
around the MTR timeout.