Bug #49931  	Incorrect type in read_log_event error Bug #49932  	mysqlbinlog max_allowed_packet hard coded to 1GB
The size limitation of ROW event is controlled by binlog-row-event-max-size (4G),
which can be larger than max-allowed-packet (1G). And the value of
max_allowed_packet is hard coded to 1GB in mysqlbinlog. But the slave
IO/SQL thread and mysqlbinlog just check the size of the ROW event by
the value of 'max_allowed_packet' when reading log event from binlog.  
So it will cause 'Event too large' or 'packet too large' error, if the
size of ROW event is bigger than the value of 'max_allowed_packet'.
      
To fix the problem, the slave IO/SQL thread and mysqlbinlog will
check the bigger one of the values of 'max_allowed_packet' and
'binlog-row-event-max-size' when reading log event from binlog. 
And add a new 'binlog-row-event-max-size' (4G by default) option
to mysqlbinlog. Then mysqlbinlog can read a large ROW event.