wl#2540 replication events checksum
Addressing `never attempt to execute' corrupted event requirement.

*This is not yet the final commit but rather an improved prototype 
 to capture all discussed arch-issues*.

TODO:

1.  extend test bases to cover few corner cases and more simulations to add.
2.  sort out if we need (A) duplication of to be content with always CRC32-ed
    FD.


LOG_EVENT_CHECKSUM_F flag is removed. Instead,
the FD event is responsible to hold the checksum alg descriptor (A).
If FD version is lower that the home of checksumming then the descriptor
will stay in an undefined value which allows to support OM -> NS scenario
(normal upgrading that is).

At relay-logging the master's FD is always examined on the matter of (A) value.
The value is propagated into FD-for-queue that affect the way the local slave
events are relay-logged (FD, Rotate, Stop). The mentioned propagation makes possible
to identify (A) correctly regardless of what server originated the event 
at time of reading RL by SQL thread.

FD is made to contain (A) as well as CRC32 signature regardless of (A) value.
That is FD starting from the home-checksumming version can be always CRC32-verified.


There are few exceptional cases.

I. The first event from master is not FD but so called fake Rotate event.
   In order to recognize its checksum status slave-master handshake is extended
   to notify the slave side about the instant value of @@global.binlog_checksum
   of the master that matches R_f checksum.

II. R_f or any real Rotate from master are relay-log right after the slave's own
    FD_s. As the last seen FD is (A)-holder for the following events stream R
    is relay-logged checksum-compatibly to FD_s.



 sql/log

    MYSQL_BIN_LOG class is extended with `relay_log_checksum_alg' memeber
                  to denote the slave's initial preference how to checksum its FD
		  as well as serves as tool to keep up FD_q's checksum_alg value
		  at time FD_q changes -> (STOP SLAVE: NULL) -> (START SLAVE: restores).


    write_cache:  fixing the end_log position because one did account systematic
                  BINLOG_CHECKSUM_LEN drift per event.


    slave side STOP event, FD, Rotate event are `flagged' right after constructor
    with appropriate value of checksum_alg.

sql/log_event

    Log_event:checksum_alg    the base class is made the placeholder to serve as
    			      mechanism to control checksumming Log_event::need_checksum().


 sql/rpl_m

   + Master_info::checksum_alg_before_fd   to hold checksum alg in use until IO thread has received FD

   + Master_info::last_master_checksum_alg  restart slave via STOP, START sequence makes IO thread to
    rotate the former session relay-log. The rotation event of the new session
    needs checksum alg of the former one.



sql/slave.cc

    get_master_version_and_clock()  elaborated slave-master handshake to
                                    make the slave be aware about R_f checksum
				    and to prepare FD_q.checksum_alg with
				    the value from the last session (if it happened)
				    or with the very first RL's FD's checksum_alg
                                    (to instruct how to checksum R_f incl one 
				     from OM).

     queue_event()   takes care of facing the first R_f, as well as exceptions
                     incl OM -> NS.


sql/sql_repl.c

     refined fake_rotate_event(), send_heartbeat_event() to deal with the correct
     checksum_alg values when prepering events.

     Providing support for SQL thread and Show-Binlog-Events handlers.